Refactoring : improve the readability of the code
