+++ package.json
@@ -5,7 +5,7 @@
   "main": "app.js",
   "scripts": {
     "test": "SET NODE_ENV=development && jest --detectOpenHandles",
-    "start": "node ./dist/src/app.js",
+    "start": "SET NODE_ENV=production && node ./dist/src/app.js",
     "dev": "SET NODE_ENV=development && nodemon ./dist/src/app.js",
     "prod": "SET NODE_ENV=production && nodemon ./dist/src/app.js",
     "build": "npm i && npm install typescript && npx tsc"
