+++ src/__tests__/assets/testData/stringTestData.ts
@@ -1,0 +1,35 @@
+export const stringWith25Char = "2i3n8sLqAe7z0m9c6R4s2W1t8";
+export const stringWith24Char = "2i3n8sLqAe7z0m9c6R4s2W18";
+export const stringWith65Char = "7xG3o9pN2fL6eK1rI0jR4tF5kG9iJ4oP0qL2mR6hA1sV9uE3vC4bR8zX5tG6a3";
+export const stringWith33Char = "9oP2qR5tF8kG1iJ4oP0qL2mR6hA1sV9uE3vC4b";
+export const emailWith5Char = "E@g.c"
+export const emailWith6Char = "a@g.us";
+export const emailWith64Char = "3Ji7lHvF0eR4wTxY2qLm5oI9pZ8kE1dG4hN6dwqwqdwwwbV9aX3c@example.com";
+export const emailWith65Char = "3sJi7lHvF0eR4wTxY2qLm5oI9pZ8kE1dG4hN6dwqwqdwwwbV9aX3c@example.com";
+export const stringWith4Char = "name";
+export const stringWith3Char = "Cha";
+export const stringWith5Char = "namee";
+export const stringWith6Char = "123456";
+export const stringWith10Char = "123456dqwdw";
+export const stringWith32Char = "9oP2qR5tF8kG1iqL2mR6hA1sV9uE3vC4";
+export const stringWith257Char = "9fxjE5pUW4y2mVvRQ6gDrMbo0GHN8PcSXKsJFWktVpvHbqnzLdta8DVIBo7FpAs2kJzlxpOw8X4sNPyCdRFMziN2eo0HQRq3rAxXvMdjbZdHXbFVYm0oYP7v4cYmBl39LHTAbh9Vd8vLjiX6YjxRHwQYnBoZJuQn0PEcLXyFfCFh1BrT2OLkC33jt9hEDZuAPX5dx9nG38eXgr1d6PSkUVRIiMf6g0C1vL8tB48eXgr1d6PSkUVRIiMf6g0C1vL8t"
+export const stringWith256Char = "9fxjE5pUW4y2mVvRQ6gDrMbo0GHN8PcSXKsJFWktVpvHbqnzLdta8DVIBo7FpAs2kJzlxpOw8X4sNPyCdRFMziN2eo0HQRq3rAxXvMdjbZdHXbFVYm0oYP7v4cYmBl39LHTAbh9Vd8vLjiX6YjxRHwQYnBoZJuQn0PEcLXyFfCFh1BrT2OLkC33jt9hEDZuAPX5dx9nG38eXgr1d6PSkUVRIiMf6g0C1vL8tB48eXgr1d6PSkUVRIiMf6g0C1vL8"
+export const stringWith62Char = "9fxjE5pUW4y2mVvRQ6gDrMbo0GHN8PcSXKsJFWktVpvHbqnzLdta8DVIBo8DVI"
+export const stringWith63Char = "9fxjE5pUW4y2mVvRQ6gDrMbo0GHN8PcSXKsJFWktVpvHbqnzLdta8DVIBo8DVIB"
+export const stringWith15Char = "9fxjE5pUW4y2mVv"
+export const stringWith16Char = "9fxjE5pUW4y2mVva"
+export const stringWith12Char = "9fxjE5pUW4y2"
+export const stringWith9Char = "9fxjE5pUW"
+export const stringWith13Char = "9fxjE5pUW4y2a"
+export const stringWith1024Char = "9fxjE5pUW4y2mVvRQ6gDrMbo0GHN8PcSXKsJFWktVpvHbqnzLdta8DVIBo7FpAs2kJzlxpOw8X4sNPyCdRFMziN2eo0HQRq3rAxXvMdjbZdHXbFVYm0oYP7v4cYmBl39LHTAbh9Vd8vLjiX6YjxRHwQYnBoZJuQn0PEcLXyFfCFh1BrT2OLkC33jt9hEDZuAPX5dx9nG38eXgr1d6PSkUVRIiMf6g0C1vL8tB49fxjE5pUW4y2mVvRQ6gDrMbo0GHN8PcSXKsJFWktVpvHbqnzLdta8DVIBo7FpAs2kJzlxpOw8X4sNPyCdRFMziN2eo0HQRq3rAxXvMdjbZdHXbFVYm0oYP7v4cYmBl39LHTAbh9Vd8vLjiX6YjxRHwQYnBoZJuQn0PEcLXyFfCFh1BrT2OLkC33jt9hEDZuAPX5dx9nG38eXgr1d6PSkUVRIiMf6g0C1vL8tB49fxjE5pUW4y2mVvRQ6gDrMbo0GHN8PcSXKsJFWktVpvHbqnzLdta8DVIBo7FpAs2kJzlxpOw8X4sNPyCdRFMziN2eo0HQRq3rAxXvMdjbZdHXbFVYm0oYP7v4cYmBl39LHTAbh9Vd8vLjiX6YjxRHwQYnBoZJuQn0PEcLXyFfCFh1BrT2OLkC33jt9hEDZuAPX5dx9nG38eXgr1d6PSkUVRIiMf6g0C1vL8tB49fxjE5pUW4y2mVvRQ6gDrMbo0GHN8PcSXKsJFWktVpvHbqnzLdta8DVIBo7FpAw8X4sNPyCdRFMziN2eo0HQRq3rAxXvMdjbZdHXbFVYm0oYP7v9fxjEvRQ6gDrMbo0GHN8PcSXKsJFWktVpvHbqnzLdta8DVIBo7FpAs2kJzlxpOw8X4sNPyCdRFMziN2eo0HQRq3rAxXvMdjbZdHXbFVYmYmBl39LHTAbh9Vd8vLjiX6YjxRHwQYnBoZJuQn0PEcLXyFfCFh1BrT2OLkC33jt9hEDZuAPX5dx9nG38eXgr1d6PSkUVRIiMf6g0C1vL8tB4C1vL8tB4Ad"
+export const stringWith1025Char = "9fxjE5pUW4y2mVvRQ6gDrMbo0GHN8PcSXKsJFWktVpvHbqnzLdta8DVIBo7FpAs2kJzlxpOw8X4sNPyCdRFMziN2eo0HQRq3rAxXvMdjbZdHXbFVYm0oYP7v4cYmBl39LHTAbh9Vd8vLjiX6YjxRHwQYnBoZJuQn0PEcLXyFfCFh1BrT2OLkC33jt9hEDZuAPX5dx9nG38eXgr1d6PSkUVRIiMf6g0C1vL8tB49fxjE5pUW4y2mVvRQ6gDrMbo0GHN8PcSXKsJFWktVpvHbqnzLdta8DVIBo7FpAs2kJzlxpOw8X4sNPyCdRFMziN2eo0HQRq3rAxXvMdjbZdHXbFVYm0oYP7v4cYmBl39LHTAbh9Vd8vLjiX6YjxRHwQYnBoZJuQn0PEcLXyFfCFh1BrT2OLkC33jt9hEDZuAPX5dx9nG38eXgr1d6PSkUVRIiMf6g0C1vL8tB49fxjE5pUW4y2mVvRQ6gDrMbo0GHN8PcSXKsJFWktVpvHbqnzLdta8DVIBo7FpAs2kJzlxpOw8X4sNPyCdRFMziN2eo0HQRq3rAxXvMdjbZdHXbFVYm0oYP7v4cYmBl39LHTAbh9Vd8vLjiX6YjxRHwQYnBoZJuQn0PEcLXyFfCFh1BrT2OLkC33jt9hEDZuAPX5dx9nG38eXgr1d6PSkUVRIiMf6g0C1vL8tB49fxjE5pUW4y2mVvRQ6gDrMbo0GHN8PcSXKsJFWktVpvHbqnzLdta8DVIBo7FpAw8X4sNPyCdRFMziN2eo0HQRq3rAxXvMdjbZdHXbFVYm0oYP7v9fxjEvRQ6gDrMbo0GHN8PcSXKsJFWktVpvHbqnzLdta8DVIBo7FpAs2kJzlxpOw8X4sNPyCdRFMziN2eo0HQRq3rAxXvMdjbZdHXbFVYmYmBl39LHTAbh9Vd8vLjiX6YjxRHwQYnBoZJuQn0PEcLXyFfCFh1BrT2OLkC33jt9hEDZuAPX5dx9nG38eXgr1d6PSkUVRIiMf6g0C1vL8tB4C1vL8tB4Ada"
+export const stringWith100Char ="9fxjE5pUW4y2mVvRQ6gDrMbo0GHN8PcSXKsJFWktVpvHbqnzLdta8DVIBo8DVIB9fxjE5pUW4y2mVvRQ6gDrMbo0GHN8PcSXKsJF"
+export const stringWith101Char = "9fxjE5pUW4y2mVvRQ6gDrMbo0GHN8PcSXKsJFWktVpvHbqnzLdta8DVIBo8DVIB9fxjE5pUW4y2mVvRQ6gDrMbo0GHN8PcSXKsJF"
+export const maxPrice = 1000;
+export const minPrice = 0;
+export const minOffer = 0.00;
+export const maxOffer = 1.00;
+export const stringWith2Char = "ab";
+export const hexWith23Char = "65e7d89b62bb29693a0d1c5"
+export const hexWith25Char = "65e7d89b62bb29693a0d1c589"
+export const hexWith24Char = "65e7d89b62bb29693a0d1c58"
